Make step arguments a concern of cucumber.
Move regexp argument matcher to a factory method on step argument.
